scripts/vps/requirements.txt
python-telegram-bot>=21.9,<22.0
python-dotenv>=1.0.0
groq>=0.5.0
